fix(process): correct TIOCSCTTY ioctl argument type
Update the `ioctl` calls used when setting the controlling terminal to
pass `TIOCSCTTY` as `libc::c_ulong`, matching the expected
platform-specific signature and avoiding type mismatch issues.

Also adjust the Cargo config lines shown in the diff by commenting
them out.

+++ dsh/src/process/pty.rs
@@ -97,7 +97,7 @@ pub fn make_controlling_terminal(slave_fd: RawFd) -> Result<()> {
 
     // Set controlling terminal
     unsafe {
-        if libc::ioctl(slave_fd, libc::TIOCSCTTY, 0) != 0 {
+        if libc::ioctl(slave_fd, libc::TIOCSCTTY as libc::c_ulong, 0) != 0 {
             return Err(anyhow::anyhow!("ioctl TIOCSCTTY failed"));
         }
     }
